Fix the reset code to not cause PPU bus conflicts
I discovered that the reset code was causing the PPU to write to CHR ROM
in a way that was likely to cause bus conflicts. Turns out, that while
the post-reset PPU ignores writes to PPUAddr, it does not ignore writes
to PPUData, which then write at whatever the unchanged address was.

This fix is to rework the way it tries to set the BG palette to black,
such that it actually waits for as long as it needs to (with a small
buffer) before trying to write data to the PPU. The target cycle count
is based on information found in the nesdev wiki re:PPU power up state.

This also improves a few other minor things, like moving the disabling
of NMI earlier (for the consoles where the PPU does not get reset).